More 64-bit uint vs size_t fixes

+++ tango/io/selector/SelectSelector.d
@@ -543,7 +543,7 @@ private class SelectSelectionSet: ISelectionSet
         _exceptionSet = exceptionSet;
     }
 
-    @property uint length()
+    @property size_t length()
     {
         return _eventCount;
     }
